+++ nodalia-cards.manifest.js
@@ -1,10 +1,10 @@
 export default {
   "pkgVersion": "1.1.0-alpha.20",
-  "contentSha256_12": "eaa90e624c88",
+  "contentSha256_12": "d2f76dd9aca6",
   "file": "nodalia-cards.bundle.js",
   "loaderFile": "nodalia-cards.js",
   "hacsFile": "nodalia-cards-1.1.0-alpha.20.js"
 };
 export const pkgVersion = "1.1.0-alpha.20";
-export const contentSha256_12 = "eaa90e624c88";
+export const contentSha256_12 = "d2f76dd9aca6";
 export const file = "nodalia-cards.bundle.js";